Workaround for shader linkage issues on Linux/Mesa
OSD requires certain shader functions to be defined by the downstream
code, including OsdPrimitiveId(), however USD is only defining these in
certain circumstances. As a result, shaders can fail to link with an
undefined symbol error.

These function definitions are now in _GetOSDCommonShaderSource() so are
defined in all cases, rather than based on the presence of certain
substrings. This may not be correct from a rendering perspective, but it
does avoid the linker error.

Only tested on Linux with Mesa so far.