Avoid needlessly copying headers into build directory
Copying the public headers into the build directory complicates
development since any edits need to be made to the copy, not the
original, which means that the changes don't show up in "git diff" and
some compiler error messages can be confusing.

Instead of copying the headers we now add ${CMAKE_SOURCE_DIR} as an
include path so any "#include <pxr/...>" operation can find the original
header in the source tree.