USDImagingGLRenderParams::clearColor is a VtValue
Back-end rendering code in UsdImagingGLEngine already accepts a VtValue
as the clear color, however this was not exposed to users via the
UsdImagingGLRenderParams structure, which instead required an actual
color vector with no possible "none" value.

By exposing the VtValue directly to callers, we allow the clear color to
be absent (rather than just black), which in turn allows multiple render
passes on Mac to composite with each other without any intermediate
buffer-clearing.

The disadvantage of a VtValue is that it isn't type safe (it is more
like an std::any than an std::optional), which means the caller could
pass in something other than a color without triggering any compilation
errors. However it does seem to be the standard idiom within USD for
passing or returning potentially empty values.